odoo_mcp/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
whoosh>=2.7.4
lxml>=5.0.0
pydantic>=2.0.0
click>=8.0.0
PyYAML>=6.0.0
python-multipart>=0.0.6
typing-extensions>=4.0.0
starlette>=0.25.0
